Refactor imports to use same pattern
