+++ apps/admin-ui/src/common/ErrorGeneric.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ErrorContainer from "common/src/components/ErrorContainer";
+import { ErrorContainer } from "common/src/components";
 import { env } from "@/env.mjs";
 import { PUBLIC_URL } from "@/common/const";
 import { Flex } from "common/styled";
