+++ apps/admin-ui/src/pages/500.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import type { GetServerSidePropsContext } from "next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
+import { ErrorContainer } from "common/src/components";
 import { getCommonServerSideProps } from "@/modules/serverUtils";
-import ErrorContainer from "common/src/components/ErrorContainer";
 
 // TODO this is a copy of 404.tsx, but with 500 instead of 404
