+++ apps/admin-ui/src/pages/_error.tsx
@@ -1,7 +1,7 @@
 import * as Sentry from "@sentry/nextjs";
 import type { NextPage } from "next";
 import Error, { type ErrorProps } from "next/error";
-import ErrorContainer from "common/src/components/ErrorContainer";
+import { ErrorContainer } from "common/src/components";
 import { env } from "@/env.mjs";
 
 const CustomErrorComponent: NextPage<ErrorProps> = (props) => {
