+++ apps/admin-ui/src/pages/application-rounds/[id]/allocation/index.tsx
@@ -40,7 +40,7 @@ import { useSearchParams } from "next/navigation";
 import { getCommonServerSideProps } from "@/modules/serverUtils";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { type GetServerSidePropsContext } from "next";
-import Error403 from "@/common/Error403";
+import { Error403 } from "@/common/Error403";
 import { createClient } from "@/common/apolloClient";
 
 const MAX_RES_UNIT_NAME_LENGTH = 35;
