+++ apps/admin-ui/src/pages/applications/[id]/index.tsx
@@ -35,12 +35,12 @@ import { BirthDate } from "@/component/BirthDate";
 import { ValueBox } from "@/component/ValueBox";
 import { TimeSelector } from "@/component/TimeSelector";
 import { getApplicantName, translateReserveeType } from "@/helpers";
-import Error404 from "@/common/Error404";
+import { Error404 } from "@/common/Error404";
 import { useCheckPermission } from "@/hooks";
 import { ApplicationDatas, Summary } from "@/styled";
 import { ApplicationStatusLabel } from "common/src/components/statuses";
 import { useDisplayError } from "common/src/hooks";
-import Error403 from "@/common/Error403";
+import { Error403 } from "@/common/Error403";
 import { AuthorizationChecker } from "@/common/AuthorizationChecker";
 import { GetServerSidePropsContext } from "next";
 import { getCommonServerSideProps } from "@/modules/serverUtils";
