+++ apps/admin-ui/src/pages/reservations/[id]/series/index.tsx
@@ -36,7 +36,7 @@ import { getReservationUrl } from "@/common/urls";
 import { getSeriesOverlapErrors } from "common/src/apolloUtils";
 import { useDisplayError } from "common/src/hooks";
 import { generateReservations } from "@/modules/generateReservations";
-import Error404 from "@/common/Error404";
+import { Error404 } from "@/common/Error404";
 import { useRouter } from "next/router";
 import { getCommonServerSideProps } from "@/modules/serverUtils";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
@@ -44,7 +44,7 @@ import { type GetServerSidePropsContext } from "next";
 import { NOT_FOUND_SSR_VALUE } from "@/common/const";
 import { hasPermission } from "@/modules/permissionHelper";
 import { useSession } from "@/hooks/auth";
-import Error403 from "@/common/Error403";
+import { Error403 } from "@/common/Error403";
 
 type NodeT = NonNullable<SeriesPageQuery["reservation"]>["reservationSeries"];
