+++ apps/ui/pages/404.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import type { GetServerSidePropsContext } from "next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { getCommonServerSideProps } from "@/modules/serverUtils";
-import ErrorContainer from "common/src/components/ErrorContainer";
+import { ErrorContainer } from "common/src/components";
 
 /// next doesn't allow getServersideProps in 404.tsx (you have to use app router for that)
 /// so all props are build time not runtime (e.g. no dynamic environment variables)
