+++ apps/ui/pages/503.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import type { GetServerSidePropsContext } from "next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { getCommonServerSideProps } from "@/modules/serverUtils";
-import ErrorContainer from "common/src/components/ErrorContainer";
+import { ErrorContainer } from "common/src/components";
 
 /// Unlike 404 and 500 this is not a standard next error page
 /// so using getServerSideProps is possible but 503 means we can't do backend calls
