+++ packages/common/src/components/index.ts
@@ -1,6 +1,7 @@
 export { default as IconButton } from "./IconButton";
-export { default as ShowAllContainer, type ShowAllContainerProps } from "./ShowAllContainer";
+export { ShowAllContainer, type ShowAllContainerProps } from "./ShowAllContainer";
 export { default as BannerNotificationsList } from "./BannerNotificationsList";
 export { default as Card } from "./Card";
 export { default as StatusLabel } from "./StatusLabel";
 export { default as Tag } from "./Tag";
+export { ErrorContainer } from "./ErrorContainer";
